requirements.txt
Flask==3.1.3
flask-cors==6.0.2
PyJWT==2.12.1
allure-pytest==2.15.3
pytest==9.0.2
pytest-xdist==3.8.0
requests==2.33.1
pyyaml==6.0.3
python-dotenv==1.0.0